+++ src/base/io/log/Tags.h
@@ -1,12 +1,6 @@
-<<<<<<< HEAD
 /* xmlcore
- * Copyright 2018-2020 SChernykh   <https://github.com/SChernykh>
- * Copyright 2016-2020 xmlcore       <https://github.com/xmlcore>, <support@xmlcore.com>
-=======
-/* XMRig
  * Copyright (c) 2018-2021 SChernykh   <https://github.com/SChernykh>
- * Copyright (c) 2016-2021 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
->>>>>>> 072881e1a1214befdd46f5823f4ba7afeb14136a
+ * Copyright (c) 2016-2021 xmlcore       <https://github.com/xmlcore>, <support@xmlcore.com>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by