+++ src/base/kernel/config/BaseTransform.cpp
@@ -1,18 +1,6 @@
-<<<<<<< HEAD
 /* xmlcore
- * Copyright 2010      Jeff Garzik <jgarzik@pobox.com>
- * Copyright 2012-2014 pooler      <pooler@litecoinpool.org>
- * Copyright 2014      Lucas Jones <https://github.com/lucasjones>
- * Copyright 2014-2016 Wolf9466    <https://github.com/OhGodAPet>
- * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
- * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
- * Copyright 2018-2020 SChernykh   <https://github.com/SChernykh>
- * Copyright 2016-2020 xmlcore       <https://github.com/xmlcore>, <support@xmlcore.com>
-=======
-/* XMRig
  * Copyright (c) 2018-2021 SChernykh   <https://github.com/SChernykh>
- * Copyright (c) 2016-2021 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
->>>>>>> 072881e1a1214befdd46f5823f4ba7afeb14136a
+ * Copyright (c) 2016-2021 xmlcore       <https://github.com/xmlcore>, <support@xmlcore.com>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -296,13 +284,9 @@ void xmlcore::BaseTransform::transformBoolean(rapidjson::Document &doc, int key,
     case IConfig::TlsKey: /* --tls */
         return add(doc, Pools::kPools, Pool::kTls, enable);
 
-<<<<<<< HEAD
-#   ifdef xmlcore_FEATURE_HTTP
-=======
     case IConfig::SubmitToOriginKey: /* --submit-to-origin */
         return add(doc, Pools::kPools, Pool::kSubmitToOrigin, enable);
-#   ifdef XMRIG_FEATURE_HTTP
->>>>>>> 072881e1a1214befdd46f5823f4ba7afeb14136a
+#   ifdef xmlcore_FEATURE_HTTP
     case IConfig::DaemonKey: /* --daemon */
         return add(doc, Pools::kPools, Pool::kDaemon, enable);
 #   endif