+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # v6.9.0
-- [#2104](https://github.com/xmrig/xmrig/pull/2104) Added [pause-on-active](https://xmrig.com/docs/miner/config/misc#pause-on-active) config option and `--pause-on-active=N` command line option.
-- [#2112](https://github.com/xmrig/xmrig/pull/2112) Added support for [Tari merge mining](https://github.com/tari-project/tari/blob/development/README.md#tari-merge-mining).
-- [#2117](https://github.com/xmrig/xmrig/pull/2117) Fixed crash when GPU mining `cn-heavy` on Zen3 system.
+- [#2104](https://github.com/xmlcore/xmlcore/pull/2104) Added [pause-on-active](https://xmlcore.com/docs/miner/config/misc#pause-on-active) config option and `--pause-on-active=N` command line option.
+- [#2112](https://github.com/xmlcore/xmlcore/pull/2112) Added support for [Tari merge mining](https://github.com/tari-project/tari/blob/development/README.md#tari-merge-mining).
+- [#2117](https://github.com/xmlcore/xmlcore/pull/2117) Fixed crash when GPU mining `cn-heavy` on Zen3 system.
 
 # v6.8.2
 - [#2080](https://github.com/xmlcore/xmlcore/pull/2080) Fixed compile error in Termux.